Minor cleanup and addition of README

+++ src/main/groovy/io/dwclark/btree/ShortRecord.java
@@ -16,18 +16,18 @@ public class ShortRecord implements Record<Short> {
     public int size() { return 4; }
 
     public Short extract(final ImmutableBytes bytes, final long pos) {
-        return Short.valueOf(bytes.readInt(pos));
+        return Short.valueOf(bytes.readShort(pos));
     }
     
     public void place(final MutableBytes bytes, final long pos, final Short val) {
-        bytes.writeInt(pos, val.intValue());
+        bytes.writeInt(pos, val.shortValue());
     }
     
     public int compare(final ImmutableBytes bytes, final long pos, final Short lhs) {
-        return Short.compare(lhs.intValue(), bytes.readInt(pos));
+        return Short.compare(lhs.shortValue(), bytes.readShort(pos));
     }
 
     public int compareInPlace(final ImmutableBytes bytes, final long pos1, final long pos2) {
-        return Short.compare(bytes.readInt(pos1), bytes.readInt(pos2));
+        return Short.compare(bytes.readShort(pos1), bytes.readShort(pos2));
     }
 }
